+++ src_mandatory/pivot.c
@@ -6,7 +6,7 @@
 /*   By: eunskim <eunskim@student.42heilbronn.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/07 22:54:59 by eunskim           #+#    #+#             */
-/*   Updated: 2023/03/12 19:28:35 by eunskim          ###   ########.fr       */
+/*   Updated: 2023/03/12 20:02:53 by eunskim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,12 +16,12 @@
 // 2. sort the copied
 // 3. get the right pivots
 
-// void	calculate_pivots(t_ps *ps, int *tmp_arr_to_sort)
+// void	calculate_pivots(t_ps *ps, size_t *tmp_arr_to_sort)
 // {
 	
 // }
 
-int	copy_stack_a(t_ps *ps, int *tmp_arr_to_sort)
+int	copy_stack_a(t_ps *ps, size_t *tmp_arr_to_sort)
 {
 	size_t	i;
 	size_t	arr_size;
@@ -38,11 +38,11 @@ int	copy_stack_a(t_ps *ps, int *tmp_arr_to_sort)
 
 int	get_pivots(t_ps *ps)
 {
-	int		*tmp_arr_to_sort;
+	size_t	*tmp_arr_to_sort;
 	size_t	arr_size;
 
 	arr_size = ps->a.max_size;
-	tmp_arr_to_sort = ft_calloc(arr_size, sizeof(int));
+	tmp_arr_to_sort = ft_calloc(arr_size, sizeof(size_t));
 	if (tmp_arr_to_sort == NULL)
 		return (EXIT_FAILURE);
 	copy_stack_a(ps, tmp_arr_to_sort);
